% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-data.r
\name{as.matrix,TableContainer-method}
\alias{as.matrix,TableContainer-method}
\alias{as.data.frame,TableContainer-method}
\title{Data Conversion}
\usage{
\S4method{as.matrix}{TableContainer}(x)

\S4method{as.data.frame}{TableContainer}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A TableContainer object.}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame. Missing values are not allowed. See \code{base::data.frame} for more details.}

\item{optional}{Logical. If \code{TRUE}, setting row names is optional. See \code{base::data.frame} for more details.}

\item{...}{additional arguments}
}
\value{
\code{as.matrix}: A matrix representation of the TableContainer object.

\code{as.data.frame}: A data frame representation of the TableContainer object.
}
\description{
Data Conversion
}
