% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Epoch.r, R/method_attributes.r,
%   R/method_bracket.r, R/method_conversion.r, R/method_dollar.r
\name{truncateTime}
\alias{truncateTime}
\alias{truncateTime,Epoch-method}
\alias{show,Epoch-method}
\alias{nrow,Epoch-method}
\alias{ncol,Epoch-method}
\alias{colnames,Epoch-method}
\alias{colnames<-,Epoch-method}
\alias{rownames,Epoch-method}
\alias{rownames<-,Epoch-method}
\alias{names,Epoch-method}
\alias{names<-,Epoch-method}
\alias{[,Epoch-method}
\alias{as.matrix,Epoch-method}
\alias{as.data.frame,Epoch-method}
\alias{$,Epoch-method}
\alias{$<-,Epoch-method}
\title{Epoch Methods}
\usage{
truncateTime(x, from, to)

\S4method{truncateTime}{Epoch}(x, from, to)

\S4method{show}{Epoch}(object)

\S4method{nrow}{Epoch}(x)

\S4method{ncol}{Epoch}(x)

\S4method{colnames}{Epoch}(x)

\S4method{colnames}{Epoch}(x) <- value

\S4method{rownames}{Epoch}(x)

\S4method{rownames}{Epoch}(x) <- value

\S4method{names}{Epoch}(x)

\S4method{names}{Epoch}(x) <- value

\S4method{[}{Epoch}(x, i, j)

\S4method{as.matrix}{Epoch}(x)

\S4method{as.data.frame}{Epoch}(x, row.names = NULL, optional = FALSE, ...)

\S4method{$}{Epoch}(x, name)

\S4method{$}{Epoch}(x, name) <- value
}
\arguments{
\item{x}{Epoch object}

\item{from}{Numeric value specifying start of new time range}

\item{to}{Numeric value specifying end of new time range}

\item{object}{Epoch object}

\item{value}{Value to set}

\item{i}{Row (electrode) indices}

\item{j}{Column (time) indices}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame. Missing values are not allowed. See \code{base::data.frame} for more details.}

\item{optional}{Logical. If \code{TRUE}, setting row names is optional. See \code{base::data.frame} for more details.}

\item{...}{additional arguments}

\item{name}{a value name, must be one of 'electrodes', 'times',
'timeRange', 'data'}
}
\value{
truncateTime: Truncated object

nrow: Number of rows in the data

ncol: Number of columns in the data

colnames: electrode names of the data

rownames: time points of the data

names: Return all available properties for an Epoch object
}
\description{
\code{truncateTime}: Truncating time range

\code{nrow}, \code{ncol}, \code{colnames}, \code{rownames}, \code{names}: Getting the data properties,
similar to base R functions.

\code{[}: Subset an Epoch object using matrix indexing syntax

\verb{$electrodes}: Get or set electrode names
\verb{$times}: Get or set time points
\verb{$timeRange}: Get time range if time points are defined
\verb{$data}: Get or set data matrix
}
