% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.r
\name{show,TableContainer-method}
\alias{show,TableContainer-method}
\title{Pretty show method for TableContainer}
\usage{
\S4method{show}{TableContainer}(object)
}
\arguments{
\item{object}{A TableContainer object.}
}
\description{
Pretty show method for TableContainer
}
