% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TableContainer.r
\docType{class}
\name{TableContainer-class}
\alias{TableContainer-class}
\alias{.TableContainer}
\title{The TableContainer Class}
\description{
A container for a matrix and associated row/column annotations. This is
similar in concept to
Bioconductor's SummarizedExperiment but with no Bioconductor dependencies.
}
\section{Slots}{

\describe{
\item{\code{table}}{A matrix, data.frame, or NULL.}

\item{\code{rowData}}{A data.frame or NULL object describing the rows.
Each row of the matrix corresponds to a row in the rowData object.}

\item{\code{colData}}{A data.frame or NULL object describing the columns.
Each column of the matrix corresponds to a row in the colData object.}

\item{\code{metaData}}{A list or NULL containing arbitrary metadata associated
with the overall data.}
}}

