% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-TableContainer.r, R/method-bracket.r,
%   R/method-data.r
\name{dim,TableContainer-method}
\alias{dim,TableContainer-method}
\alias{dimnames,TableContainer-method}
\alias{nrow,TableContainer-method}
\alias{ncol,TableContainer-method}
\alias{[.TableContainer}
\alias{tblData}
\alias{tblData<-}
\alias{rowData}
\alias{rowData<-}
\alias{colData}
\alias{colData<-}
\alias{metadata}
\alias{metadata<-}
\title{Container Methods}
\usage{
\S4method{dim}{TableContainer}(x)

\S4method{dimnames}{TableContainer}(x)

\S4method{nrow}{TableContainer}(x)

\S4method{ncol}{TableContainer}(x)

\method{[}{TableContainer}(x, i, j, ...)

tblData(object)

tblData(object) <- value

rowData(object)

rowData(object) <- value

colData(object)

colData(object) <- value

metadata(object)

metadata(object) <- value
}
\arguments{
\item{x}{A TableContainer object.}

\item{i}{Row indices for subsetting. If only \code{i} is provided, it will return the entire row(s).}

\item{j}{Column indices for subsetting.}

\item{...}{Additional arguments.}

\item{object}{A TableContainer object.}

\item{value}{A matrix, data.frame, or NULL.}
}
\value{
\code{dim}, \code{dimnames}, \code{nrow}, and \code{ncol}: the respective dimensions, dimnames, number of rows, and number of columns of the TableContainer object. When the table slot is NULL, the dimensions are derived from the rowData and colData slots.

\code{[}: A new TableContainer object with the selected data.

\code{tblData}, \code{rowData}, \code{colData}, and \code{metadata}: the respective slots of the TableContainer object.

\verb{tblData<-}, \verb{rowData<-}, \verb{colData<-}, and \verb{metadata<-}: update the respective slots and return the modified object.
}
\description{
Container Methods
}
