% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.r
\name{.printTable}
\alias{.printTable}
\alias{.printMeta}
\alias{show,TableContainer-method}
\title{Show method for TableContainer and its components}
\usage{
.printTable(tbl, header = "")

.printMeta(meta, name)

\S4method{show}{TableContainer}(object)
}
\arguments{
\item{tbl}{A table-like object (e.g., matrix, data.frame).}

\item{header}{A string to print as the first element in the column names row.}

\item{meta}{A list of metadata items.}

\item{name}{A string representing the name of the metadata.}

\item{object}{A TableContainer object.}
}
\description{
The show method uses \code{.printTable} to display the table and \code{.printMeta} to display the rowMeta, colMeta, and metaData.
}
